Break up NodeWidgetModal into smaller components
