+++ src/core/components/modals/NodeWidget/index.ts
@@ -1,0 +1,3 @@
+import NodeWidgetModal from './NodeWidget';
+
+export default NodeWidgetModal;
